requirements.txt
click==8.1.3
openai==0.27.2
pandas==1.5.3
pandas_stubs==1.2.0.58
python-dotenv==1.0.0
setuptools==58.0.4
tiktoken==0.3.3
